requirements.txt
jieba==0.42.1
jsonlines==3.1.0
matplotlib==3.3.0
numpy==1.21.6
pandas==1.3.5
Pillow==9.2.0
pylcs==0.0.7
pytorch_lightning==1.7.7
torch==1.12.1
tqdm==4.64.1
transformers==4.22.2
undecorated==0.3.0
